requirements.txt
Pillow==8.2.0
pandas==1.1.3
plotly==5.1.0
streamlit==0.83.0
regex==2021.8.3
networkx==2.6.2
matplotlib==3.4.2
nltk==3.6.2
emoji==1.4.2
